Update my code 3

+++ basic_math.py
@@ -42,6 +42,6 @@ def get_median(number_list):
     if length % 2:
         median = number_list[length / 2]
     else:
-        median = number_list[length / 2 - 1] + number_list[length / 2] / 2
+        median = int(number_list[length / 2 - 1] + number_list[length / 2] / 2)
 
     return median
